Move limelight enums and records to Constants.java and Structures.java respectively
