+++ src/main/java/com/spartronics4915/frc2025/subsystems/vision/LimelightDevice.java
@@ -4,9 +4,12 @@ import java.util.ArrayList;
 import java.util.Optional;
 
 import com.spartronics4915.frc2025.Constants.VisionConstants;
+import com.spartronics4915.frc2025.Constants.VisionConstants.LimelightModel;
+import com.spartronics4915.frc2025.Constants.VisionConstants.LimelightRole;
 import com.spartronics4915.frc2025.LimelightHelpers;
 import com.spartronics4915.frc2025.LimelightHelpers.RawFiducial;
 import com.spartronics4915.frc2025.util.Structures.LimelightConstants;
+import com.spartronics4915.frc2025.util.Structures.VisionMeasurement;
 
 import edu.wpi.first.math.geometry.Pose2d;
 import edu.wpi.first.net.PortForwarder;
@@ -15,16 +18,6 @@ import edu.wpi.first.wpilibj2.command.SubsystemBase;
 import swervelib.SwerveDrive;
 
 public class LimelightDevice extends SubsystemBase {
-
-    public static record VisionMeasurement(Pose2d pose, double timestamp) {}
-
-    public enum LimelightModel {
-        LIMELIGHT_3, LIMELIGHT_3G
-    }
-
-    public enum LimelightRole {
-        NOTHING, REEF, STATION, OBSERVER
-    }
 
     private final String name;
     private final LimelightModel model;
@@ -77,7 +70,7 @@ public class LimelightDevice extends SubsystemBase {
         if (rejectUpdate) {
             return Optional.empty();
         } else {
-            return Optional.of(new VisionMeasurement(megaTag2.pose, megaTag2.timestampSeconds));
+            return Optional.of(new VisionMeasurement(megaTag2.pose, megaTag2.timestampSeconds, role));
         }
     }
